[fed] Split up federated test CMake file.
- Collect all federated test files into the same directory.
- Independently list the files.
