+++ tests/cpp/CMakeLists.txt
@@ -1,3 +1,6 @@
+# The testxgboost executable is created in the top level CMakeLists. Most of the
+# properties and compilation flags are already set. We just need to add source files and
+# link gtest here.
 if(USE_DMLC_GTEST)
   if(NOT TARGET gtest)
     message(FATAL_ERROR "USE_DMLC_GTEST=ON but dmlc-core didn't bundle gtest")
@@ -6,12 +9,17 @@ if(USE_DMLC_GTEST)
 else()
   find_package(GTest REQUIRED)
 endif()
+
 file(GLOB_RECURSE TEST_SOURCES "*.cc")
 
 if(USE_CUDA)
   file(GLOB_RECURSE CUDA_TEST_SOURCES "*.cu")
   list(APPEND TEST_SOURCES ${CUDA_TEST_SOURCES})
 endif()
+
+# We will add them back later to separate the definition.
+file(GLOB_RECURSE FEDERATED_TEST_SOURCES "plugin/federated/*.*")
+list(REMOVE_ITEM TEST_SOURCES ${FEDERATED_TEST_SOURCES})
 
 file(GLOB_RECURSE SYCL_TEST_SOURCES "plugin/test_sycl_*.cc")
 list(REMOVE_ITEM TEST_SOURCES ${SYCL_TEST_SOURCES})
@@ -48,14 +56,14 @@ if(PLUGIN_SYCL)
 endif()
 
 if(PLUGIN_FEDERATED)
-  target_include_directories(testxgboost PRIVATE ${xgboost_SOURCE_DIR}/plugin/federated)
-  target_link_libraries(testxgboost PRIVATE federated_client)
-else()
-  file(GLOB_RECURSE FEDERATED_TEST_SOURCES "plugin/*_federated_*.*")
-  list(REMOVE_ITEM TEST_SOURCES ${FEDERATED_TEST_SOURCES})
+  add_subdirectory(${xgboost_SOURCE_DIR}/tests/cpp/plugin/federated)
 endif()
 
-target_sources(testxgboost PRIVATE ${TEST_SOURCES} ${xgboost_SOURCE_DIR}/plugin/example/custom_obj.cc)
+target_sources(
+  testxgboost PRIVATE
+  ${TEST_SOURCES}
+  ${xgboost_SOURCE_DIR}/plugin/example/custom_obj.cc
+)
 
 if(USE_CUDA AND PLUGIN_RMM)
   target_include_directories(testxgboost PRIVATE ${CUDA_INCLUDE_DIRS})
@@ -63,7 +71,6 @@ endif()
 
 target_include_directories(testxgboost
   PRIVATE
-  ${GTEST_INCLUDE_DIRS}
   ${xgboost_SOURCE_DIR}/include
   ${xgboost_SOURCE_DIR}/dmlc-core/include)
 target_link_libraries(testxgboost
